refactor: assembling the app features in modular way (#9129)
