+++ api/dify_app.py
@@ -1,0 +1,5 @@
+from flask import Flask
+
+
+class DifyApp(Flask):
+    pass
