+++ api/extensions/ext_code_based_extension.py
@@ -1,7 +1,8 @@
 from core.extension.extension import Extension
+from dify_app import DifyApp
 
 
-def init():
+def init_app(app: DifyApp):
     code_based_extension.init()
