+++ api/extensions/ext_compress.py
@@ -1,11 +1,13 @@
-from flask import Flask
-
 from configs import dify_config
+from dify_app import DifyApp
 
 
-def init_app(app: Flask):
-    if dify_config.API_COMPRESSION_ENABLED:
-        from flask_compress import Compress
+def is_enabled() -> bool:
+    return dify_config.API_COMPRESSION_ENABLED
 
-        compress = Compress()
-        compress.init_app(app)
+
+def init_app(app: DifyApp):
+    from flask_compress import Compress
+
+    compress = Compress()
+    compress.init_app(app)
