+++ api/extensions/ext_hosting_provider.py
@@ -1,9 +1,10 @@
-from flask import Flask
-
 from core.hosting_configuration import HostingConfiguration
 
 hosting_configuration = HostingConfiguration()
 
 
-def init_app(app: Flask):
+from dify_app import DifyApp
+
+
+def init_app(app: DifyApp):
     hosting_configuration.init_app(app)
