+++ api/extensions/ext_migrate.py
@@ -1,5 +1,9 @@
-import flask_migrate
+from dify_app import DifyApp
 
 
-def init(app, db):
+def init_app(app: DifyApp):
+    import flask_migrate
+
+    from extensions.ext_database import db
+
     flask_migrate.Migrate(app, db)
