+++ api/extensions/ext_proxy_fix.py
@@ -1,9 +1,8 @@
-from flask import Flask
-
 from configs import dify_config
+from dify_app import DifyApp
 
 
-def init_app(app: Flask):
+def init_app(app: DifyApp):
     if dify_config.RESPECT_XFORWARD_HEADERS_ENABLED:
         from werkzeug.middleware.proxy_fix import ProxyFix
