+++ api/extensions/ext_set_secretkey.py
@@ -1,0 +1,6 @@
+from configs import dify_config
+from dify_app import DifyApp
+
+
+def init_app(app: DifyApp):
+    app.secret_key = dify_config.SECRET_KEY
