+++ api/extensions/ext_warnings.py
@@ -1,0 +1,7 @@
+from dify_app import DifyApp
+
+
+def init_app(app: DifyApp):
+    import warnings
+
+    warnings.simplefilter("ignore", ResourceWarning)
